+++ components/navbar/ProfileDropdown.vue
@@ -77,7 +77,7 @@
         aria-role="listitem"
         :value="lang.value"
         :class="{ 'is-active': $i18n.locale === lang.value }"
-        @click="setUserLocale(lang.value)"
+        @click="usePreferencesStore().setUserLocale(lang.value)"
       >
         <span>{{ lang.flag }} {{ lang.label }}</span>
       </NeoDropdownItem>
@@ -88,10 +88,8 @@
 <script lang="ts" setup>
 import { NeoDropdown, NeoDropdownItem, NeoIcon } from '@kodadot1/brick'
 import { useIdentityStore } from '@/stores/identity'
-import {
-  langsFlags as langsFlagsList,
-  setUserLocale,
-} from '@/utils/config/i18n'
+import { langsFlags as langsFlagsList, usePreferencesStore } from '@/stores/preferences'
+
 import { openConnectWalletModal } from '@/components/common/ConnectWallet/useConnectWallet'
 import ConnectWalletButton from '@/components/shared/ConnectWalletButton.vue'
 import { useProfileOnboardingStore } from '@/stores/profileOnboarding'
